+++ tests/test_plots_1d.py
@@ -4,7 +4,7 @@ matplotlib.use('Agg')  # Use a non-GUI backend
 import pytest
 from uadapy.dr import uamds
 import uadapy.data as data
-from uadapy.plotting import plots1D
+from uadapy.plotting import plots_1d
 
 @pytest.fixture
 def sample_distribution():
@@ -23,59 +23,59 @@ def sample_distributions():
 @pytest.mark.mpl_image_compare(baseline_dir="baseline")
 def test_generate_boxplot(sample_distribution):
     """Test generate_boxplot function."""
-    fig, axs = plots1D.generate_boxplot(sample_distribution)
+    fig, axs = plots_1d.generate_boxplot(sample_distribution)
     return fig
 
 @pytest.mark.mpl_image_compare(baseline_dir="baseline")
 def test_generate_violinplot(sample_distribution):
     """Test generate_violinplot function."""
-    fig, axs = plots1D.generate_violinplot(sample_distribution)
+    fig, axs = plots_1d.generate_violinplot(sample_distribution)
     return fig
 
 @pytest.mark.mpl_image_compare(baseline_dir="baseline")
 def test_generate_dotplot(sample_distribution):
     """Test generate_dotplot function."""
-    fig, axs = plots1D.generate_dotplot(sample_distribution)
+    fig, axs = plots_1d.generate_dotplot(sample_distribution)
     return fig
 
 @pytest.mark.mpl_image_compare(baseline_dir="baseline")
 def test_generate_stripplot(sample_distribution):
     """Test generate_stripplot function."""
-    fig, axs = plots1D.generate_stripplot(sample_distribution)
+    fig, axs = plots_1d.generate_stripplot(sample_distribution)
     return fig
 
 @pytest.mark.mpl_image_compare(baseline_dir="baseline")
 def test_generate_swarmplot(sample_distribution):
     """Test generate_swarmplot function."""
-    fig, axs = plots1D.generate_swarmplot(sample_distribution)
+    fig, axs = plots_1d.generate_swarmplot(sample_distribution)
     return fig
 
 @pytest.mark.mpl_image_compare(baseline_dir="baseline")
 def test_generate_multidim_boxplot(sample_distributions):
     """Test generate_multidim_boxplot function."""
-    fig, axs = plots1D.generate_multidim_boxplot(sample_distributions)
+    fig, axs = plots_1d.generate_multidim_boxplot(sample_distributions)
     return fig
 
 @pytest.mark.mpl_image_compare(baseline_dir="baseline")
 def test_generate_multidim_violinplot(sample_distributions):
     """Test generate_multidim_violinplot function."""
-    fig, axs = plots1D.generate_multidim_violinplot(sample_distributions)
+    fig, axs = plots_1d.generate_multidim_violinplot(sample_distributions)
     return fig
 
 @pytest.mark.mpl_image_compare(baseline_dir="baseline")
 def test_generate_multidim_dotplot(sample_distributions):
     """Test generate_multidim_dotplot function."""
-    fig, axs = plots1D.generate_multidim_dotplot(sample_distributions)
+    fig, axs = plots_1d.generate_multidim_dotplot(sample_distributions)
     return fig
 
 @pytest.mark.mpl_image_compare(baseline_dir="baseline")
 def test_generate_multidim_stripplot(sample_distributions):
     """Test generate_multidim_stripplot function."""
-    fig, axs = plots1D.generate_multidim_stripplot(sample_distributions)
+    fig, axs = plots_1d.generate_multidim_stripplot(sample_distributions)
     return fig
 
 @pytest.mark.mpl_image_compare(baseline_dir="baseline")
 def test_generate_multidim_swarmplot(sample_distributions):
     """Test generate_multidim_swarmplot function."""
-    fig, axs = plots1D.generate_multidim_swarmplot(sample_distributions)
+    fig, axs = plots_1d.generate_multidim_swarmplot(sample_distributions)
     return fig
